feat: Refactor Dagger Hilt modules and update locale usage
This commit refactors the Dagger Hilt dependency injection setup by splitting the `SingletonModule` into more focused modules: `RepositoriesDI`, `Network_API_DI`, and `AndroidDI`.

- `RepositoriesDI`: Provides dependencies related to data repositories, Room database, DAOs, and `DataCache`.
- `Network_API_DI`: Provides dependencies for network operations, including Retrofit instances for `FootballApi` and `NewsApi`, and `NetworkConnectivityReceiver`.
- `AndroidDI`: Renamed from `SingletonModule`, this module now provides Android-specific dependencies like `AlarmManager`, `SharedPreferences`, and `Context`.

Additionally, the method for getting the current locale in `MatchDetailsActivity.kt` has been updated from `getResources().configuration.getLocales().get(0)` to `getResources().configuration.locales.get(0)` to align with newer Android API versions.

Minor layout adjustments were made in `activity_match_details.xml` by changing `layout_height` from `wrap_content` or `match_parent` to `0dp` with a `layout_weight` of `2` for `TextView` elements within vertical `LinearLayouts` to improve their distribution.
