+++ app/src/main/res/layout/activity_match_details.xml
@@ -103,7 +103,8 @@
                 android:orientation="vertical">
                 <TextView
                     android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
+                    android:layout_height="0dp"
+                    android:layout_weight="2"
                     android:text="@{matchDetails.refereeName}"
                     android:drawablePadding="5dp"
                     app:drawableLeftCompat="@drawable/football_whistle_of_referee" />
@@ -114,7 +115,8 @@
 
                 <TextView
                     android:layout_width="match_parent"
-                    android:layout_height="match_parent"
+                    android:layout_height="0dp"
+                    android:layout_weight="2"
                     android:gravity="center_vertical"
                     android:text="@{matchDetails.arenaName}"
                     android:drawablePadding="5dp"
@@ -127,7 +129,8 @@
                 <TextView
                     android:id="@+id/tv_channels"
                     android:layout_width="match_parent"
-                    android:layout_height="match_parent"
+                    android:layout_height="0dp"
+                    android:layout_weight="2"
                     android:gravity="center_vertical"
                     android:drawablePadding="5dp"
                     app:drawableLeftCompat="@drawable/tv" />
